refactor(runtime): split orchestrator/tests — extract history (final slice)
Wraps the orchestrator/tests split. Pull the last 10 basic
history-threading and tool-call accumulation tests out of
`tests/mod.rs` into a new `tests/history.rs`:

- `first_turn_sends_only_current_message`
- `second_turn_includes_prior_history`
- `current_message_not_duplicated`
- `seeded_history_included_in_llm_call`
- `three_turns_accumulate_history`
- `different_conversations_are_isolated`
- `single_tool_call_adds_observation_to_next_request`
- `two_tool_calls_handled_in_single_iteration`
- `two_tool_calls_both_observations_sent_to_llm`
- `three_tool_calls_handled_in_single_iteration`

The `ollama_tool_calls` and `ollama_tool_calls_with_args` helpers
lift up into the top-of-file helpers block alongside `ollama_answer`
since they're shared with sibling modules.

`tests/mod.rs` is now pure helpers (`ollama_*`, `build`, `mount_answer`,
`messages_in`, `MockExtTool`) + `mod` declarations — no test bodies
remain. The original 3536-line monolith is now distributed across:

  mod.rs       ~210   helpers + declarations
  bus_worker.rs ~530   12 tests
  end_turn.rs   ~470    6 tests
  history.rs    ~450   10 tests  ← new
  misc.rs       ~360    8 tests
  multimodal.rs ~500   15 tests
  sanitize.rs   ~260    9 tests
  streaming.rs  ~330    5 tests
  subagent.rs   ~600   10 tests
                ────  ── ──
  9 files     ~3650   75 tests

cargo test -p assistant-runtime --lib reports 220 passed (same as
before).
